+++ pkg/backup/create_remote.go
@@ -6,7 +6,7 @@ import (
 	"github.com/AlexAkulov/clickhouse-backup/pkg/status"
 )
 
-func (b *Backuper) CreateToRemote(backupName, diffFrom, diffFromRemote, tablePattern string, partitions []string, schemaOnly, rbac, backupConfig, resume bool, version string, commandId int) error {
+func (b *Backuper) CreateToRemote(backupName, diffFrom, diffFromRemote, tablePattern string, partitions []string, schemaOnly, rbac, backupConfig, skipCheckPartsColumns, resume bool, version string, commandId int) error {
 	ctx, cancel, err := status.Current.GetContextWithCancel(commandId)
 	if err != nil {
 		return err
@@ -16,7 +16,7 @@ func (b *Backuper) CreateToRemote(backupName, diffFrom, diffFromRemote, tablePat
 	if backupName == "" {
 		backupName = NewBackupName()
 	}
-	if err := b.CreateBackup(backupName, tablePattern, partitions, schemaOnly, rbac, backupConfig, version, commandId); err != nil {
+	if err := b.CreateBackup(backupName, tablePattern, partitions, schemaOnly, rbac, backupConfig, skipCheckPartsColumns, version, commandId); err != nil {
 		return err
 	}
 	if err := b.Upload(backupName, diffFrom, diffFromRemote, tablePattern, partitions, schemaOnly, resume, commandId); err != nil {